netlify/functions/match_details/requirements.txt
fotmob-wrapper
aiohttp
requests
